+++ framework/validators/ExistValidator.php
@@ -9,6 +9,8 @@ namespace yii\validators;
 
 use Yii;
 use yii\base\InvalidConfigException;
+use yii\base\Model;
+use yii\db\ActiveRecord;
 
 /**
  * ExistValidator validates that the attribute value exists in a table.
@@ -137,14 +139,28 @@ class ExistValidator extends Validator
             if ($this->allowArray) {
                 throw new InvalidConfigException('The "targetAttribute" property must be configured as a string.');
             }
-            $params = [];
+            $conditions = [];
             foreach ($targetAttribute as $k => $v) {
-                $params[$v] = is_int($k) ? $model->$v : $model->$k;
+                $conditions[$v] = is_int($k) ? $model->$v : $model->$k;
             }
         } else {
-            $params = [$targetAttribute => $model->$attribute];
+            $conditions = [$targetAttribute => $model->$attribute];
         }
-        return $params;
+
+        if (!$model instanceof ActiveRecord) {
+            return $conditions;
+        }
+
+        return $this->prefixConditions($model, $conditions);
+    }
+
+    /**
+     * @param Model $model the data model to be validated
+     * @return string Target class name
+     */
+    private function getTargetClass($model)
+    {
+        return $this->targetClass === null ? get_class($model) : $this->targetClass;
     }
 
     /**
@@ -188,5 +204,29 @@ class ExistValidator extends Validator
         }
 
         return $query;
+    }
+
+    /**
+     * Prefix conditions with aliases
+     *
+     * @param ActiveRecord $model
+     * @param array $conditions
+     * @return array
+     */
+    private function prefixConditions($model, $conditions)
+    {
+        $targetModelClass = $this->getTargetClass($model);
+
+        /** @var ActiveRecord $targetModelClass */
+        $query = $targetModelClass::find();
+        $tableAliases = array_keys($query->getTablesUsedInFrom());
+        $primaryTableAlias = $tableAliases[0];
+        $prefixedConditions = [];
+        foreach ($conditions as $columnName => $columnValue) {
+            $prefixedColumn = "{{{$primaryTableAlias}}}.[[{$columnName}]]";
+            $prefixedConditions[$prefixedColumn] = $columnValue;
+        }
+
+        return $prefixedConditions;
     }
 }